requirements.txt
Django==1.9.8
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
djangorestframework-filters==0.9.0